Move menu_input_dialog_keyboard_display to retroarch.c
